Update download script for yaml grammar to remove subdir
schema.generated.cc file was put into a subdirectory in #47 to make the
build work on macOs.

While the solution works great overall, it may cause problems for
other tooling (e.g. bazel).

This commit updates download script with merging both schema.generated.cc
and scanner.cc into one file. It solves the build problem and avoids a
directory without go files.

Closes #62.
